requirements.txt
openai
gradio>=5.5.0
fireworks-ai